requirements.txt
streamlit
agno
sympy
pydantic>=2.0,<3
duckduckgo-search
sentence-transformers
chonkie[semantic]
huggingface-hub
google-genai
easyocr
groq
openai
Pillow
numpy
ddgs
python-dotenv
wikipedia
lancedb
sqlalchemy
sqlite-utils
aiofiles
asyncio
altair